backend/auth/requirements.txt
psycopg2-binary>=2.9.0
pyjwt>=2.8.0
